position_adjust_text->adjust_text parameter
